remove ping command. format.
ping is not necessary. you can keep a plugin alive longer by changing
the garbage collector configs. here is information from the nushell
docs:

---

Plugin Garbage Collector

As mentioned above, Nu comes with a plugin garbage collector which
automatically stops plugins that are not actively in use after a period
of time (by default, 10 seconds). This behavior is fully configurable:

$env.config.plugin_gc = {
    # Settings for plugins not otherwise specified:
    default: {
        enabled: true # set to false to never automatically stop plugins
        stop_after: 10sec # how long to wait after the plugin is inactive before stopping it
    }
    # Settings for specific plugins, by plugin name
    # (i.e. what you see in `plugin list`):
    plugins: {
        gstat: {
            stop_after: 1min
        }
        inc: {
            stop_after: 0sec # stop as soon as possible
        }
        example: {
            enabled: false # never stop automatically
        }
    }
}