experimental reconfigure directories
after researching package management options, it is my opinion that the
best package manager for nushell scripts and plugins is npm.

although nushell is written in rust, cargo (the package manager for
rust), is not appropriate for nushell scripts because it installs
packages in a random hash-based folder that is unpredictable by the
nushell scripts. however, npm installs packages in a predictable
path located at `node_modules/` in the current working directory, which
makes loding a nushell script trivial. npm can also be used with
binaries, and it supports optional, environment-dependent dependencies,
making it possible to include pre-built binaries in the package manager
available on all platforms.

nu_plugin_torch can still be distributed via cargo, but it should also
be made available on npm, like all other nushell scripts and plugins.